requirements.txt
streamlit==1.46.1
python-dotenv==1.1.1
langchain==0.3.26
langchain-core==0.3.84
langchain-community==0.3.27
langchain-openai==0.2.14
SQLAlchemy==2.0.41
pyodbc